Move 'symbolic' package up in module tree for easier refactor
