+++ src/python/zquantum/core/wip/circuit/conversions/qiskit_conversions.py
@@ -4,9 +4,9 @@ from typing import Tuple, List, Union
 
 import qiskit
 
-from .symbolic.qiskit_expressions import expression_from_qiskit, QISKIT_DIALECT
-from .symbolic.sympy_expressions import expression_from_sympy, SYMPY_DIALECT
-from .symbolic.translations import translate_expression
+from ..symbolic.qiskit_expressions import expression_from_qiskit, QISKIT_DIALECT
+from ..symbolic.sympy_expressions import expression_from_sympy, SYMPY_DIALECT
+from ..symbolic.translations import translate_expression
 from .. import (
     X,
     Y,
